Add a utility to identify the position of a dictionary item in Monaco Editor (#94)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This PR adds a utility function for finding a dictionary item by ID in a Monaco Editor value. This uses the `parse` from `vscode-html-languageservice` to traverse the DOM structure and is consolidated with the `DataDictionary` which is then reinterpreted into columns and line numbers for the Monaco Editors position.

### 🎫 Issues

<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->
Addresses the first bullet in #53

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
Check to make sure the logic and comments are clear. This may not be the most efficient solution as it traverses the complete tree until it finds the dictionary item, however optimizations may come later from the use of this function in the `MonacoAdapter` class.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [x] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->
- Integrate this into the `MonacoAdapter`
- Test this using `MonacoAdapterAction`