Fix line number and column number which should start at 1 (#105)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
An assumption was made that line number and column number start at 0, this is incorrect as both start at 1. This PR updates the code to reflect this.

### 🎫 Issues

<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->
Related to #53

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
Check the tests for accuracy, the dictionary ID should be identified once past the "<" of the HTML tag it refers to, otherwise it should revert to the parent. Additionally the position should be updated for the line number and column change as stated in the description above.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.