Update the Monaco Adapter service to include the ability to update the position inside the monaco value (#106)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This adds a callback for the config supplied to Monaco Adapter actions. This will allow a position update so that Navigation messages may trigger a new update from within the Monaco Editor and scroll into view.

### 🎫 Issues

<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->
Related to #53

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.