Add wasm-pack to the github workflow and a test script for the wasm package (#128)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This change should add the wasm package to the build gate and prevent errors from being pushed. The package is currently private and requires `wasm-pack` to run.

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
The build gate should pass and include the tests in the wasm package. To run locally, follow the steps in the CONTRIBUTING.md file for the wasm package.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [x] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.