Remove the postinstall script as it causes errors during installation as a dependency (#129)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
The `postinstall` script is firing when the fast-tooling-react package is added as a dependency. To prevent this only the `prepare` script is utilized. It is unknown at this time why the change was made during migration.

Edit: well apparently it was because fast-tooling-react installs first even though fast-tooling is a dependency. Still looking for a solution.

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
Run installation locally to see if any issues arise.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.