Update message system to send all state data for outgoing messages (#161)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This PR updates all outgoing messages to include the "stateful" data contained in the web worker. This should help performance on the client side when assessing incoming messages from the message system and reduce complexity.

The reason this work has been started is due to the upcoming undo/redo or history messages, which make it difficult to remain in sync when storing incoming messages as history.

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->

Note that this change is going into a feature branch. The next steps should also do the same.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [x] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->
- Update web components message system handling in accordance with these changes
- Update react components message system handling in accordance with these changes