Simplify the handling of messages sent to the Form and Navigation components (#165)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This change simplifies the handling messages sent from the `MessageSystem`. It is going into the feature branch for the outgoing message updates. This should not change any behavior, but simplify how messages are being handled.

### 🎫 Issues

<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->
Closes #162

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.