Allow multiple outgoing messages to be sent (#180)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This change makes it so that all outgoing messages are an array of messages. This will allow for a single message update to create multiple messages that are then sent out. As this is dealt with internally, there are no breaking changes.

This change also adds a manual testing section for the new undo/redo alpha released shortcuts.

### 🎫 Issues

<!---
List and link relevant issues here using the keyword "closes"
if this PR will close an issue, eg. closes #411
-->

Resolves #179

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
This should now enable, in the manual testing app, seeing "delete" shortcut action followed by "undo".

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [x] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->
- More robust testing and implementation work is needed to ensure previous and next (undo/redo) maneuvering through the history stack does not cause errors to throw