Fixed an issue in the HTML render layer web component (#188)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
The previous fix addressed one issue with custom messages in the HTML Render, but there was a further attempt at processing custom messages on the HTML Render Layer. Additionally, the HTML Render was attempting to set the dataDictionary when a custom message was sent so the logic to understand when a message is one that needs attention has been moved to the parent if block.

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
This was checked using `npm link`. It appears to work correctly with no more runtime errors.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [ ] I have updated the project documentation to reflect my changes.