Add scaffolding for documentation content (#204)
# Pull Request

## 📖 Description

<!--- Provide some background and a description of your work. -->
This PR adds blank pages (and some style enhancements) to the documentation website. These are to be filled out one at a time after this PR is complete. The scaffolding should give an idea of the approach for how a user might visit the documentation site and interact with it.

Reliant on https://github.com/microsoft/fast-tooling/pull/203

## 👩‍💻 Reviewer Notes

<!---
Provide some notes for reviewers to help them provide targeted feedback and testing.
-->
Note that this change does not update the generated HTML in the `docs/` folder, this will only be done via a workflow dispatch.

## ✅ Checklist

### General

<!--- Review the list and put an x in the boxes that apply. -->

- [ ] I have added tests for my changes.
- [x] I have tested my changes.
- [x] I have updated the project documentation to reflect my changes.

## ⏭ Next Steps

<!---
If there is relevant follow-up work to this PR, please list any existing issues or provide brief descriptions of what you would like to do next.
-->
- Create workflow for generating documentation and committing it to the repository